requirements.txt
Flask
python-dotenv
mongoengine
openai
